Refactor to limit exported symbols and use new Identity API
This commit moves major parts of the workflow engine into an internal
package in order to limit the number of symbols that are exported from
the module.

The commit also changes the Identity client to use the new Identity
api from the servicesdk.
