--- api/call.go
@@ -1,7 +1,0 @@
-package api
-
-type Call interface {
-	Step
-
-	CalledStep() Step
-}
