--- api/loader.go
@@ -1,10 +1,0 @@
-package api
-
-import "github.com/lyraproj/pcore/px"
-
-var LyraLinkPath = px.PathType(`lyralink`)
-var GoPluginPath = px.PathType(`goplugin`)
-var PpManifestPath = px.PathType(`ppmanifest`)
-var YamlManifestPath = px.PathType(`yamlmanifest`)
-
-const LyraDlvConfigKey = `Lyra::DlvConfig`
