--- api/workflow.go
@@ -1,7 +1,0 @@
-package api
-
-type Workflow interface {
-	Step
-
-	Steps() []Step
-}
