+++ go.mod
@@ -3,7 +3,11 @@ module github.com/lyraproj/wfe
 require (
 	github.com/hashicorp/go-hclog v0.8.0
 	github.com/lyraproj/issue v0.0.0-20190606092846-e082d6813d15
-	github.com/lyraproj/pcore v0.0.0-20190614121034-381c91502dac
-	github.com/lyraproj/servicesdk v0.0.0-20190617121125-a22cb98187cb
+	github.com/lyraproj/pcore v0.0.0-20190618142417-30605b6ee043
+	github.com/lyraproj/servicesdk v0.0.0-20190618142858-870593a059dc
 	gonum.org/v1/gonum v0.0.0-20190331200053-3d26580ed485
 )
+
+replace github.com/lyraproj/pcore => github.com/thallgren/pcore v0.0.0-20190619151240-bebc8c351bb4
+
+replace github.com/lyraproj/servicesdk => github.com/thallgren/servicesdk v0.0.0-20190619152445-7481da553aae
