+++ internal/action.go
@@ -1,15 +1,15 @@
-package wfe
+package internal
 
 import (
 	"github.com/hashicorp/go-hclog"
 	"github.com/lyraproj/pcore/px"
 	"github.com/lyraproj/pcore/types"
 	"github.com/lyraproj/servicesdk/serviceapi"
-	"github.com/lyraproj/wfe/api"
+	"github.com/lyraproj/wfe/wfe"
 )
 
 type action struct {
-	Step
+	step
 	api px.ObjectType
 }
 
@@ -19,23 +19,23 @@ var expectedType = types.NewCallableType(
 	types.NewTupleType([]px.Type{ioType}, nil), ioType, nil)
 */
 
-func Action(def serviceapi.Definition) api.Step {
+func newAction(def serviceapi.Definition) wfe.Step {
 	a := &action{}
 	a.Init(def)
 	return a
 }
 
-func (s *action) Init(d serviceapi.Definition) {
-	s.Step.Init(d)
+func (a *action) Init(d serviceapi.Definition) {
+	a.step.initStep(d)
 	if i, ok := d.Properties().Get4(`interface`); ok {
-		s.api = i.(px.ObjectType)
+		a.api = i.(px.ObjectType)
 	}
 }
 
-func (s *action) Run(ctx px.Context, parameters px.OrderedMap) px.OrderedMap {
-	service := s.GetService(ctx)
-	hclog.Default().Debug(`executing action`, `name`, s.name)
-	result := service.Invoke(ctx, s.Name(), `do`, parameters)
+func (a *action) Run(ctx px.Context, parameters px.OrderedMap) px.OrderedMap {
+	service := a.GetService(ctx)
+	hclog.Default().Debug(`executing action`, `name`, a.name)
+	result := service.Invoke(ctx, a.Name(), `do`, parameters)
 	if m, ok := result.(px.OrderedMap); ok {
 		return m
 	}
@@ -45,19 +45,19 @@ func (s *action) Run(ctx px.Context, parameters px.OrderedMap) px.OrderedMap {
 	panic(result.String())
 }
 
-func (s *action) Label() string {
-	return StepLabel(s)
+func (a *action) Label() string {
+	return StepLabel(a)
 }
 
 func (a *action) Identifier() string {
 	return StepId(a)
 }
 
-func (s *action) Style() string {
+func (a *action) Style() string {
 	return `action`
 }
 
-func (a *action) WithIndex(index int) api.Step {
+func (a *action) WithIndex(index int) wfe.Step {
 	ac := *a // Copy by value
 	ac.setIndex(index)
 	return &ac
