--- service/crud_test.go
@@ -1,7 +1,0 @@
-package service_test
-
-import "testing"
-
-func TestDeferredDelete(t *testing.T) {
-
-}
